add NO_I2C_ROBOT build flag + env variable
This fixes a compile error I've been getting on all my machines (arch and ubuntu). I'm guessing that it's due to the i2c headers having changed at some point.

In any case, we only need to build in arduino support (that is, local i2c stuff) on the robot, so it doesn't matter if other machines don't have this capability.

Now, if the NO_I2C_ROBOT environment variable is set, the arduino i2c parts of the beerobot code are not built (though one could still connect to a "Nor-bot" remotely)